🔧 Refactor and consolidate cafeteria and sales-related types and components
- Moved Productos_Elaboraciones and TarjetasVentas types to cafeteria/types
- Removed ventas-cafeteria directory and redistributed its contents
- Updated import paths across components to reflect new type locations
- Added `disponible` property to TarjetasVentas type
- Simplified type imports and removed redundant type definitions
