--- app/(with-layout)/(almacen-cafeteria)/ventas-cafeteria/actions.ts
@@ -1,92 +1,0 @@
-'use server';
-
-import { ElaboracionesSchema, VentasCafeteriaSchema } from '@/lib/schemas';
-import { revalidateTag } from 'next/cache';
-import { cookies } from 'next/headers';
-import { InferInput } from 'valibot';
-
-export async function addVentaCafeteria(
-  venta: InferInput<typeof VentasCafeteriaSchema>
-): Promise<{ data: string | null; error: string | null }> {
-  const token = cookies().get('session')?.value || null;
-  const res = await fetch(process.env.BACKEND_URL_V2 + '/cafeteria/ventas/', {
-    method: 'POST',
-    headers: {
-      Authorization: `Bearer ${token}`,
-    },
-    body: JSON.stringify(venta),
-  });
-  if (!res.ok) {
-    if (res.status === 401)
-      return {
-        data: null,
-        error: 'No autorizado',
-      };
-    if (res.status === 404)
-      return {
-        data: null,
-        error: 'Error al introducir los datos.',
-      };
-    if (res.status === 400) {
-      const data = await res.json();
-      return {
-        data: null,
-        error: data.detail,
-      };
-    }
-    return {
-      data: null,
-      error: 'Algo salió mal.',
-    };
-  }
-  revalidateTag('ventas-cafeteria');
-  return {
-    data: 'Venta agregada con éxito.',
-    error: null,
-  };
-}
-
-export async function deleteVentaCafeteria({
-  id,
-}: {
-  id: number;
-}): Promise<{ data: string | null; error: string | null }> {
-  const token = cookies().get('session')?.value || null;
-  const res = await fetch(
-    process.env.BACKEND_URL_V2 + '/cafeteria/ventas/' + id + '/',
-    {
-      method: 'DELETE',
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    }
-  );
-  if (!res.ok) {
-    if (res.status === 401)
-      return {
-        data: null,
-        error: 'No autorizado',
-      };
-    if (res.status === 404)
-      return {
-        data: null,
-        error: 'Venta no encontrada.',
-      };
-    if (res.status === 400) {
-      const data = await res.json();
-      return {
-        data: null,
-        error: data.detail,
-      };
-    }
-    return {
-      data: null,
-      error: 'Algo salió mal.',
-    };
-  }
-  revalidateTag('ventas-cafeteria');
-  return {
-    data: 'Venta eliminada con éxito.',
-    error: null,
-  };
-}
